refact: SearchBar component created and fix result search responsivity
