+++ src/pages/home/home.js
@@ -1,4 +1,4 @@
-import Header from "../../components/header/header";
+import Header from "../../components/Header/Header";
 import TrendingBox from "../../components/TrendingBox/TrendingBox";
 
 export default function Home(){
